+++ tests/test_misc_tools.py
@@ -13,15 +13,3 @@ def test_show_logfile():
         assert "threedi-qgis-log.txt" in message
 
     show_logfile_action.on_unload()  # Doesn't do anything, used for coverage.
-
-
-def test_about():
-    iface = mock.Mock()
-    show_about_action = misc_tools.About(iface)
-    with mock.patch.object(misc_tools, "pop_up_info") as mock_pop_up_info:
-        show_about_action.run()
-        args, kwargs = mock_pop_up_info.call_args
-        message = args[0]
-        assert "version" in message
-
-    show_about_action.on_unload()  # Doesn't do anything, used for coverage.