+++ tool_statistics/presets.py
@@ -204,7 +204,7 @@ TS_REDUCTION_ANALYSIS_PRESETS = Preset(
     "Replacing these flowlines with orifices may speed up the "
     "simulation "
     "without large impact on the results. Import the highlighted lines "
-    "from the aggregation result into your 3Di spatialite as "
+    "from the aggregation result into your Rana schematisation geopackage as "
     "'ts_reducers' and use this query to replace line elements ("
     "example "
     "for pipe):\n\n"
@@ -296,10 +296,10 @@ water_on_street_aggregations_1d2d = [
 WATER_ON_STREET_DURATION_0D1D_PRESET = Preset(
     name="Manhole: Water on street duration (0D1D)",
     description="Time [s] that the water level in manholes exceeds the drain level.\n\n"
-                "In 3Di models without 2D, this is the level at which water flows onto the street (i.e., where the "
+                "In Rana models without 2D, this is the level at which water flows onto the street (i.e., where the "
                 "storage area changes from what is specified at the connection node to what is specified as manhole "
                 "storage area in the global settings).\n\n"
-                "⚠ Do not use this preset for 3Di models with 2D. In such models, the drain level defined at the "
+                "⚠ Do not use this preset for Rana models with 2D. In such models, the drain level defined at the "
                 "manhole is not always the level at which water flows onto the street. If the drain level is lower "
                 "than the bottom level (lowest pixel) of the 2D cell the manhole is in, the water must rise to the "
                 "2D cell's bottom level before it can flow onto the street.",
@@ -313,12 +313,12 @@ WATER_ON_STREET_DURATION_0D1D_PRESET = Preset(
 WATER_ON_STREET_DURATION_1D2D_PRESET = Preset(
     name="Manhole: Water on street duration (1D2D)",
     description="Time [s] that the water level in manholes exceeds the 1D2D exchange level.\n\n"
-                "In 3Di models with 2D, this is the level at which water flows onto the street. The exchange level is "
+                "In Rana models with 2D, this is the level at which water flows onto the street. The exchange level is "
                 "the maximum of two values: the drain level specified for the manhole, or the bottom level (lowest "
                 "pixel) of the 2D cell the manhole is in.\n\n"
                 "⚠ Manholes that have no connection to the 2D domain do not have an exchange level. The 'water on "
                 "street duration' is always 0 for these manholes.\n\n"
-                "⚠ Do not use this preset for 3Di models without 2D. In such models, none of the manholes have a "
+                "⚠ Do not use this preset for Rana models without 2D. In such models, none of the manholes have a "
                 "connection to the 2D domain, so the 'water on street duration' will be 0 for all manholes.",
     aggregations=water_on_street_aggregations_1d2d,
     nodes_style=STYLE_WATER_ON_STREET_DURATION_NODE,
@@ -338,10 +338,10 @@ max_depth_on_street_aggregations = [
 MAX_DEPTH_ON_STREET_0D1D_PRESETS = Preset(
     name="Manhole: Max water depth on street (0D1D)",
     description="Maximum water depth on manholes, calculated as maximum water level - drain level\n\n"
-                "In 3Di models without 2D, this is the level at which water flows onto the street (i.e., where the "
+                "In Rana models without 2D, this is the level at which water flows onto the street (i.e., where the "
                 "storage area changes from what is specified at the connection node to what is specified as manhole "
                 "storage area in the global settings).\n\n"
-                "⚠ Do not use this preset for 3Di models with 2D. In such models, the drain level defined at the "
+                "⚠ Do not use this preset for Rana models with 2D. In such models, the drain level defined at the "
                 "manhole is not always the level at which water flows onto the street. If the drain level is lower "
                 "than the bottom level (lowest pixel) of the 2D cell the manhole is in, the water must rise to the "
                 "2D cell's bottom level before it can flow onto the street.",
@@ -355,12 +355,12 @@ MAX_DEPTH_ON_STREET_0D1D_PRESETS = Preset(
 MAX_DEPTH_ON_STREET_1D2D_PRESETS = Preset(
     name="Manhole: Max water depth on street (1D2D)",
     description="Maximum water depth on manholes, calculated as maximum water level - 1D2D exchange level. \n\n"
-                "In 3Di models with 2D, this is the level at which water flows onto the street. The exchange level is "
+                "In Rana models with 2D, this is the level at which water flows onto the street. The exchange level is "
                 "the maximum of two values: the drain level specified for the manhole, or the bottom level (lowest "
                 "pixel) of the 2D cell the manhole is in.\n\n"
                 "⚠ Manholes that have no connection to the 2D domain do not have an exchange level. The 'water depth "
                 "on street' is NULL for these manholes.\n\n"
-                "⚠ Do not use this preset for 3Di models without 2D. In such models, none of the manholes have a "
+                "⚠ Do not use this preset for Rana models without 2D. In such models, none of the manholes have a "
                 "connection to the 2D domain, so the 'water depth on street' will be NULL for all manholes.",
     aggregations=max_depth_on_street_aggregations,
     nodes_style=STYLE_MANHOLE_WATER_DEPTH_1D2D_NODE,
@@ -382,10 +382,10 @@ MIN_FREEBOARD_0D1D_PRESETS = Preset(
     name="Manhole: Minimum freeboard (0D1D)",
     description="Minimum freeboard for manholes, "
                 "i.e. the difference between the maximum water level and the manhole drain level.\n\n"
-                "In 3Di models without 2D, this is the level at which water flows onto the street (i.e., where the "
+                "In Rana models without 2D, this is the level at which water flows onto the street (i.e., where the "
                 "storage area changes from what is specified at the connection node to what is specified as manhole "
                 "storage area in the global settings).\n\n"
-                "⚠ Do not use this preset for 3Di models with 2D. In such models, the drain level defined at the "
+                "⚠ Do not use this preset for Rana models with 2D. In such models, the drain level defined at the "
                 "manhole is not always the level at which water flows onto the street. If the drain level is lower "
                 "than the bottom level (lowest pixel) of the 2D cell the manhole is in, the water must rise to the "
                 "2D cell's bottom level before it can flow onto the street.",
@@ -400,12 +400,12 @@ MIN_FREEBOARD_1D2D_PRESETS = Preset(
     name="Manhole: Minimum freeboard (1D2D)",
     description="Minimum freeboard for manholes, "
                 "i.e. the difference between the maximum water level and the 1D2D exchange level.\n\n"
-                "In 3Di models with 2D, this is the level at which water flows onto the street. The exchange level is "
+                "In Rana models with 2D, this is the level at which water flows onto the street. The exchange level is "
                 "the maximum of two values: the drain level specified for the manhole, or the bottom level (lowest "
                 "pixel) of the 2D cell the manhole is in.\n\n"
                 "⚠ Manholes that have no connection to the 2D domain do not have an exchange level. The 'minimum "
                 "freeboard' is always NULL for these manholes.\n\n"
-                "⚠ Do not use this preset for 3Di models without 2D. In such models, none of the manholes have a "
+                "⚠ Do not use this preset for Rana models without 2D. In such models, none of the manholes have a "
                 "connection to the 2D domain, so the 'minimum freeboard' will be NULL for all manholes.",
     aggregations=max_depth_on_street_aggregations,
     nodes_style=STYLE_MANHOLE_MIN_FREEBOARD_1D2D,